refactor: 편의시설 검색 관련 코드 리팩토링
