+++ src/routes/map/methods/searchPlaces.tsx
@@ -1,14 +1,15 @@
 import { CategoryCode } from "@/utils/types";
-import { removeMarkers, displayPlaces } from "./placeService";
+import { removeMarkers, displayPlaces } from "./renderPlaces";
+
+import React from 'react'
 
 export const searchPlaces = (
-  ps: kakao.maps.services.Places | undefined,
+  ps: kakao.maps.services.Places,
   map: kakao.maps.Map | null,
   selectedCategory: "" | CategoryCode,
   markers: React.MutableRefObject<kakao.maps.Marker[]>,
   customOverlayRef: React.MutableRefObject<kakao.maps.CustomOverlay | null>
 ) => {
-  if (!ps || !map || !selectedCategory) return;
   ps.categorySearch(selectedCategory, (data, status) => {
     if (status !== window.kakao.maps.services.Status.ERROR) {
       removeMarkers(markers);
